야믈에 리이슈토큰 api/auth까지만, 로그아웃에 접근권한 추가, 서비스에 리이슈토큰이 쿠키가없어도 안터지게,

+++ src/main/java/com/deliveryinsider/domain/auth/services/AuthService.java
@@ -184,20 +184,25 @@ public class AuthService {
                 );
 
         // 회원 조회
-        Long userId = Long.valueOf(
+        if (refreshToken != null) {
+            Long userId = Long.valueOf(
                 jwtProvider.extractClaims(refreshToken).getSubject()
-        );
-
-        User user = authMapper.findById(userId);
-
-        if (user == null) {
-            throw new NotRegisteredException(
-                    "존재하지 않는 회원입니다."
             );
-        }
 
-        // DB의 Refresh Token 삭제
-        authMapper.deleteRefreshToken(user.getId());
+
+            User user = authMapper.findById(userId);
+
+            if (user == null) {
+                throw new NotRegisteredException(
+                    "존재하지 않는 회원입니다."
+                );
+            }
+
+            if (user != null) {
+                // DB의 Refresh Token 삭제
+                authMapper.deleteRefreshToken(user.getId());
+            }
+        }
 
         // Cookie 삭제
         cookieManager.deleteCookie(
@@ -264,29 +269,5 @@ public class AuthService {
                 .email(savedUser.getEmail())
                 .build();
     }
-    @Transactional
-    public void logout(
-        HttpServletRequest request,
-        HttpServletResponse response
-    ) {
-        String refreshToken = jwtProvider
-            .extractRefreshToken(request)
-            .orElse(null);
-
-        if (refreshToken != null) {
-            User user = authMapper.findByRefreshToken(refreshToken);
-
-            if (user != null) {
-                authMapper.updateRefreshToken(user.getId(), null);
-            }
-        }
-
-        cookieManager.setCookie(
-            response,
-            jwtConfig.refreshTokenCookieName(),
-            "",
-            0,
-            jwtConfig.reissUri()
-        );
-    }
+   
 }
